cep_library/requirements.txt
psutil==5.9.1
fastapi==0.95.1
networkx==2.8.4
numpy==1.22.4
opencv_python==4.7.0.72
pydantic==1.10.6
pymongo==4.3.3
Pympler==1.0.1
requests==2.28.2
pandas==2.1.0
uvicorn==0.22.0
ortools==9.12.4544
paho-mqtt==2.1.0
redis==5.0.1
python-multipart==0.0.9
matplotlib==3.8.2
librosa==0.10.1
